+++ src/Symfony/Bridge/Twig/Tests/Extension/WorkflowExtensionTest.php
@@ -14,6 +14,7 @@ namespace Symfony\Bridge\Twig\Tests\Extension;
 use PHPUnit\Framework\TestCase;
 use Symfony\Bridge\Twig\Extension\WorkflowExtension;
 use Symfony\Component\Workflow\Definition;
+use Symfony\Component\Workflow\MarkingStore\MethodMarkingStore;
 use Symfony\Component\Workflow\Metadata\InMemoryMetadataStore;
 use Symfony\Component\Workflow\Registry;
 use Symfony\Component\Workflow\SupportStrategy\ClassInstanceSupportStrategy;
@@ -49,21 +50,20 @@ class WorkflowExtensionTest extends TestCase
             );
         }
         $definition = new Definition($places, $transitions, null, $metadataStore);
-        $workflow = new Workflow($definition);
+        $workflow = new Workflow($definition, new MethodMarkingStore());
 
         $registry = new Registry();
         $addWorkflow = method_exists($registry, 'addWorkflow') ? 'addWorkflow' : 'add';
         $supportStrategy = class_exists(InstanceOfSupportStrategy::class)
-            ? new InstanceOfSupportStrategy(\stdClass::class)
-            : new ClassInstanceSupportStrategy(\stdClass::class);
+            ? new InstanceOfSupportStrategy(Subject::class)
+            : new ClassInstanceSupportStrategy(Subject::class);
         $registry->$addWorkflow($workflow, $supportStrategy);
         $this->extension = new WorkflowExtension($registry);
     }
 
     public function testCanTransition()
     {
-        $subject = new \stdClass();
-        $subject->marking = [];
+        $subject = new Subject();
 
         $this->assertTrue($this->extension->canTransition($subject, 't1'));
         $this->assertFalse($this->extension->canTransition($subject, 't2'));
@@ -71,8 +71,7 @@ class WorkflowExtensionTest extends TestCase
 
     public function testGetEnabledTransitions()
     {
-        $subject = new \stdClass();
-        $subject->marking = [];
+        $subject = new Subject();
 
         $transitions = $this->extension->getEnabledTransitions($subject);
 
@@ -83,9 +82,7 @@ class WorkflowExtensionTest extends TestCase
 
     public function testHasMarkedPlace()
     {
-        $subject = new \stdClass();
-        $subject->marking = [];
-        $subject->marking = ['ordered' => 1, 'waiting_for_payment' => 1];
+        $subject = new Subject(['ordered' => 1, 'waiting_for_payment' => 1]);
 
         $this->assertTrue($this->extension->hasMarkedPlace($subject, 'ordered'));
         $this->assertTrue($this->extension->hasMarkedPlace($subject, 'waiting_for_payment'));
@@ -94,12 +91,10 @@ class WorkflowExtensionTest extends TestCase
 
     public function testGetMarkedPlaces()
     {
-        $subject = new \stdClass();
-        $subject->marking = [];
-        $subject->marking = ['ordered' => 1, 'waiting_for_payment' => 1];
+        $subject = new Subject(['ordered' => 1, 'waiting_for_payment' => 1]);
 
         $this->assertSame(['ordered', 'waiting_for_payment'], $this->extension->getMarkedPlaces($subject));
-        $this->assertSame($subject->marking, $this->extension->getMarkedPlaces($subject, false));
+        $this->assertSame($subject->getMarking(), $this->extension->getMarkedPlaces($subject, false));
     }
 
     public function testGetMetadata()
@@ -107,13 +102,32 @@ class WorkflowExtensionTest extends TestCase
         if (!class_exists(InMemoryMetadataStore::class)) {
             $this->markTestSkipped('This test requires symfony/workflow:4.1.');
         }
-        $subject = new \stdClass();
-        $subject->marking = [];
+        $subject = new Subject();
 
         $this->assertSame('workflow title', $this->extension->getMetadata($subject, 'title'));
         $this->assertSame('ordered title', $this->extension->getMetadata($subject, 'title', 'orderer'));
         $this->assertSame('t1 title', $this->extension->getMetadata($subject, 'title', $this->t1));
         $this->assertNull($this->extension->getMetadata($subject, 'not found'));
         $this->assertNull($this->extension->getMetadata($subject, 'not found', $this->t1));
+    }
+}
+
+final class Subject
+{
+    private $marking;
+
+    public function __construct($marking = null)
+    {
+        $this->marking = $marking;
+    }
+
+    public function getMarking()
+    {
+        return $this->marking;
+    }
+
+    public function setMarking($marking)
+    {
+        $this->marking = $marking;
     }
 }