+++ src/Symfony/Component/Workflow/Tests/EventListener/GuardListenerTest.php
@@ -13,6 +13,7 @@ use Symfony\Component\Workflow\EventListener\ExpressionLanguage;
 use Symfony\Component\Workflow\EventListener\GuardExpression;
 use Symfony\Component\Workflow\EventListener\GuardListener;
 use Symfony\Component\Workflow\Marking;
+use Symfony\Component\Workflow\Tests\Subject;
 use Symfony\Component\Workflow\Transition;
 use Symfony\Component\Workflow\WorkflowInterface;
 
@@ -130,13 +131,12 @@ class GuardListenerTest extends TestCase
 
     private function createEvent(Transition $transition = null)
     {
-        $subject = new \stdClass();
-        $subject->marking = new Marking();
+        $subject = new Subject();
         $transition = $transition ?: new Transition('name', 'from', 'to');
 
         $workflow = $this->getMockBuilder(WorkflowInterface::class)->getMock();
 
-        return new GuardEvent($subject, $subject->marking, $transition, $workflow);
+        return new GuardEvent($subject, new Marking($subject->getMarking() ?? []), $transition, $workflow);
     }
 
     private function configureAuthenticationChecker($isUsed, $granted = true)