+++ src/Symfony/Component/Workflow/Tests/Subject.php
@@ -1,0 +1,23 @@
+<?php
+
+namespace Symfony\Component\Workflow\Tests;
+
+final class Subject
+{
+    private $marking;
+
+    public function __construct($marking = null)
+    {
+        $this->marking = $marking;
+    }
+
+    public function getMarking()
+    {
+        return $this->marking;
+    }
+
+    public function setMarking($marking)
+    {
+        $this->marking = $marking;
+    }
+}